+++ app/Http/Controllers/UserController.php
@@ -29,7 +29,6 @@ use App\Http\Models\Verify;
 use App\Mail\activeUser;
 use App\Mail\newTicket;
 use App\Mail\replyTicket;
-use App\Mail\resetPassword;
 use Illuminate\Http\Request;
 use Redirect;
 use Response;
@@ -531,14 +530,6 @@ class UserController extends Controller
         return Response::view('user.invite', $view);
     }
 
-    // 公开的邀请码列表
-    public function free(Request $request)
-    {
-        $view['inviteList'] = Invite::query()->where('uid', 0)->where('status', 0)->paginate();
-
-        return Response::view('user.free', $view);
-    }
-
     // 生成邀请码
     public function makeInvite(Request $request)
     {
@@ -557,274 +548,6 @@ class UserController extends Controller
         $obj->save();
 
         return Response::json(['status' => 'success', 'data' => '', 'message' => '生成成功']);
-    }
-
-    // 激活账号页
-    public function activeUser(Request $request)
-    {
-        if ($request->method() == 'POST') {
-            $username = trim($request->get('username'));
-
-            // 是否开启账号激活
-            if (!self::$systemConfig['is_active_register']) {
-                Session::flash('errorMsg', '系统未开启账号激活功能，请联系管理员');
-
-                return Redirect::back()->withInput();
-            }
-
-            // 查找账号
-            $user = User::query()->where('username', $username)->first();
-            if (!$user) {
-                Session::flash('errorMsg', '账号不存在，请重试');
-
-                return Redirect::back();
-            } elseif ($user->status < 0) {
-                Session::flash('errorMsg', '账号已禁止登陆，无需激活');
-
-                return Redirect::back();
-            } elseif ($user->status > 0) {
-                Session::flash('errorMsg', '账号无需激活');
-
-                return Redirect::back();
-            }
-
-            // 24小时内激活次数限制
-            $activeTimes = 0;
-            if (Cache::has('activeUser_' . md5($username))) {
-                $activeTimes = Cache::get('activeUser_' . md5($username));
-                if ($activeTimes >= self::$systemConfig['active_times']) {
-                    Session::flash('errorMsg', '同一个账号24小时内只能请求激活' . self::$systemConfig['active_times'] . '次，请勿频繁操作');
-
-                    return Redirect::back();
-                }
-            }
-
-            // 生成激活账号的地址
-            $token = md5(self::$systemConfig['website_name'] . $username . microtime());
-            $verify = new Verify();
-            $verify->type = 1;
-            $verify->user_id = $user->id;
-            $verify->token = $token;
-            $verify->status = 0;
-            $verify->save();
-
-            // 发送邮件
-            $activeUserUrl = self::$systemConfig['website_url'] . '/active/' . $token;
-            $title = '重新激活账号';
-            $content = '请求地址：' . $activeUserUrl;
-
-            try {
-                Mail::to($user->username)->send(new activeUser(self::$systemConfig['website_name'], $activeUserUrl));
-                Helpers::addEmailLog($user->id, $title, $content);
-            } catch (\Exception $e) {
-                Helpers::addEmailLog($user->id, $title, $content, 0, $e->getMessage());
-            }
-
-            Cache::put('activeUser_' . md5($username), $activeTimes + 1, 1440);
-            Session::flash('successMsg', '激活邮件已发送，如未收到，请查看垃圾邮箱');
-
-            return Redirect::back();
-        } else {
-            return Response::view('user.activeUser');
-        }
-    }
-
-    // 激活账号
-    public function active(Request $request, $token)
-    {
-        if (empty($token)) {
-            return Redirect::to('login');
-        }
-
-        $verify = Verify::query()->where('token', $token)->with('user')->first();
-        if (empty($verify)) {
-            return Redirect::to('login');
-        } elseif (empty($verify->user)) {
-            Session::flash('errorMsg', '该链接已失效');
-
-            return Response::view('user.active');
-        } elseif ($verify->status == 1) {
-            Session::flash('errorMsg', '该链接已失效');
-
-            return Response::view('user.active');
-        } elseif ($verify->user->status != 0) {
-            Session::flash('errorMsg', '该账号无需激活.');
-
-            return Response::view('user.active');
-        } elseif (time() - strtotime($verify->created_at) >= 1800) {
-            Session::flash('errorMsg', '该链接已过期');
-
-            // 置为已失效
-            $verify->status = 2;
-            $verify->save();
-
-            return Response::view('user.active');
-        }
-
-        // 更新账号状态
-        $ret = User::query()->where('id', $verify->user_id)->update(['status' => 1]);
-        if (!$ret) {
-            Session::flash('errorMsg', '账号激活失败');
-
-            return Redirect::back();
-        }
-
-        // 置为已使用
-        $verify->status = 1;
-        $verify->save();
-
-        // 账号激活后给邀请人送流量
-        if ($verify->user->referral_uid) {
-            $transfer_enable = self::$systemConfig['referral_traffic'] * 1048576;
-
-            User::query()->where('id', $verify->user->referral_uid)->increment('transfer_enable', $transfer_enable);
-            User::query()->where('id', $verify->user->referral_uid)->update(['enable' => 1]);
-        }
-
-        Session::flash('successMsg', '账号激活成功');
-
-        return Response::view('user.active');
-    }
-
-    // 重设密码页
-    public function resetPassword(Request $request)
-    {
-        if ($request->method() == 'POST') {
-            $username = trim($request->get('username'));
-
-            // 是否开启重设密码
-            if (!self::$systemConfig['is_reset_password']) {
-                Session::flash('errorMsg', '系统未开启重置密码功能，请联系管理员');
-
-                return Redirect::back()->withInput();
-            }
-
-            // 查找账号
-            $user = User::query()->where('username', $username)->first();
-            if (!$user) {
-                Session::flash('errorMsg', '账号不存在，请重试');
-
-                return Redirect::back();
-            }
-
-            // 24小时内重设密码次数限制
-            $resetTimes = 0;
-            if (Cache::has('resetPassword_' . md5($username))) {
-                $resetTimes = Cache::get('resetPassword_' . md5($username));
-                if ($resetTimes >= self::$systemConfig['reset_password_times']) {
-                    Session::flash('errorMsg', '同一个账号24小时内只能重设密码' . self::$systemConfig['reset_password_times'] . '次，请勿频繁操作');
-
-                    return Redirect::back();
-                }
-            }
-
-            // 生成取回密码的地址
-            $token = md5(self::$systemConfig['website_name'] . $username . microtime());
-            $verify = new Verify();
-            $verify->type = 1;
-            $verify->user_id = $user->id;
-            $verify->token = $token;
-            $verify->status = 0;
-            $verify->save();
-
-            // 发送邮件
-            $resetPasswordUrl = self::$systemConfig['website_url'] . '/reset/' . $token;
-            $title = '重置密码';
-            $content = '请求地址：' . $resetPasswordUrl;
-
-            try {
-                Mail::to($user->username)->send(new resetPassword(self::$systemConfig['website_name'], $resetPasswordUrl));
-                Helpers::addEmailLog($user->id, $title, $content);
-            } catch (\Exception $e) {
-                Helpers::addEmailLog($user->id, $title, $content, 0, $e->getMessage());
-            }
-
-            Cache::put('resetPassword_' . md5($username), $resetTimes + 1, 1440);
-            Session::flash('successMsg', '重置成功，请查看邮箱');
-
-            return Redirect::back();
-        } else {
-            return Response::view('user.resetPassword');
-        }
-    }
-
-    // 重设密码
-    public function reset(Request $request, $token)
-    {
-        if ($request->method() == 'POST') {
-            $password = trim($request->get('password'));
-            $repassword = trim($request->get('repassword'));
-
-            if (empty($token)) {
-                return Redirect::to('login');
-            } elseif (empty($password) || empty($repassword)) {
-                Session::flash('errorMsg', '密码不能为空');
-
-                return Redirect::back();
-            } elseif (md5($password) != md5($repassword)) {
-                Session::flash('errorMsg', '两次输入密码不一致，请重新输入');
-
-                return Redirect::back();
-            }
-
-            // 校验账号
-            $verify = Verify::query()->where('token', $token)->with('User')->first();
-            if (empty($verify)) {
-                return Redirect::to('login');
-            } elseif ($verify->status == 1) {
-                Session::flash('errorMsg', '该链接已失效');
-
-                return Redirect::back();
-            } elseif ($verify->user->status < 0) {
-                Session::flash('errorMsg', '账号已被禁用');
-
-                return Redirect::back();
-            } elseif (Hash::check($password, $verify->user->password)) {
-                Session::flash('errorMsg', '新旧密码一样，请重新输入');
-
-                return Redirect::back();
-            }
-
-            // 更新密码
-            $ret = User::query()->where('id', $verify->user_id)->update(['password' => Hash::make($password)]);
-            if (!$ret) {
-                Session::flash('errorMsg', '重设密码失败');
-
-                return Redirect::back();
-            }
-
-            // 置为已使用
-            $verify->status = 1;
-            $verify->save();
-
-            Session::flash('successMsg', '新密码设置成功，请自行登录');
-
-            return Redirect::back();
-        } else {
-            if (empty($token)) {
-                return Redirect::to('login');
-            }
-
-            $verify = Verify::query()->where('token', $token)->with('user')->first();
-            if (empty($verify)) {
-                return Redirect::to('login');
-            } elseif (time() - strtotime($verify->created_at) >= 1800) {
-                Session::flash('errorMsg', '该链接已过期');
-
-                // 置为已失效
-                $verify->status = 2;
-                $verify->save();
-
-                // 重新获取一遍verify
-                $view['verify'] = Verify::query()->where('token', $token)->with('user')->first();
-
-                return Response::view('user.reset', $view);
-            }
-
-            $view['verify'] = $verify;
-
-            return Response::view('user.reset', $view);
-        }
     }
 
     // 使用优惠券
@@ -1253,13 +976,5 @@ class UserController extends Controller
 
             return Response::json(['status' => 'fail', 'data' => '', 'message' => '充值失败']);
         }
-    }
-
-    // 切换语言
-    public function switchLang(Request $request, $locale)
-    {
-        Session::put("locale", $locale);
-
-        return Redirect::back();
     }
 }
