+++ resources/views/auth/active.blade.php
