+++ resources/views/auth/register.blade.php
