+++ resources/views/auth/reset.blade.php
