+++ resources/views/auth/resetPassword.blade.php
